+++ ydb/src/client_coordination/session/mod.rs
@@ -1,8 +1,8 @@
 pub mod acquire_options;
+pub mod controller;
 pub mod create_options;
 pub mod describe_options;
 pub mod lease;
 pub mod session;
 pub mod session_options;
 pub mod update_options;
-
