+++ ydb/src/grpc_wrapper/raw_coordination_service/session/describe_semaphore.rs
@@ -4,7 +4,7 @@ use ydb_grpc::ydb_proto::{
 };
 
 use crate::{
-    client_coordination::{list_types::SemaphoreDescription, session::session::IdentifiedMessage},
+    client_coordination::session::controller::IdentifiedMessage,
     grpc_wrapper::{grpc::proto_issues_to_ydb_issues, raw_errors::RawError},
     WatchMode, YdbStatusError,
 };
@@ -23,6 +23,54 @@ pub(crate) struct RawDescribeSemaphoreResult {
     pub req_id: u64,
     pub semaphore_description: SemaphoreDescription,
     pub watch_added: bool,
+}
+
+#[derive(Debug, serde::Serialize)]
+#[allow(dead_code)]
+pub struct SemaphoreSession {
+    pub order_id: u64,
+    pub session_id: u64,
+    pub timeout_millis: u64,
+    pub count: u64,
+    pub data: Vec<u8>,
+}
+
+#[derive(Debug, serde::Serialize)]
+#[allow(dead_code)]
+pub struct SemaphoreDescription {
+    pub name: String,
+    pub data: Vec<u8>,
+    pub count: u64,
+    pub limit: u64,
+    pub ephemeral: bool,
+    pub owners: Vec<SemaphoreSession>,
+    pub waiters: Vec<SemaphoreSession>,
+}
+
+impl From<ydb_grpc::ydb_proto::coordination::SemaphoreSession> for SemaphoreSession {
+    fn from(value: ydb_grpc::ydb_proto::coordination::SemaphoreSession) -> Self {
+        Self {
+            order_id: value.order_id,
+            session_id: value.session_id,
+            timeout_millis: value.timeout_millis,
+            count: value.count,
+            data: value.data,
+        }
+    }
+}
+
+impl From<ydb_grpc::ydb_proto::coordination::SemaphoreDescription> for SemaphoreDescription {
+    fn from(value: ydb_grpc::ydb_proto::coordination::SemaphoreDescription) -> Self {
+        Self {
+            name: value.name,
+            data: value.data,
+            count: value.count,
+            limit: value.limit,
+            ephemeral: value.ephemeral,
+            owners: value.owners.into_iter().map(|v| v.into()).collect(),
+            waiters: value.waiters.into_iter().map(|v| v.into()).collect(),
+        }
+    }
 }
 
 impl RawDescribeSemaphoreRequest {
